+++ src/database/models/comment/mod.rs
@@ -1,7 +1,9 @@
 pub mod comment;
 pub mod entity;
 pub mod votes;
+pub mod forms;
 
 pub use entity::*;
 pub use comment::*;
 pub use votes::*;
+pub use forms::*;